RAG-main/requirements.txt
# ===========================================
# RAG Document Chat - Dependencies
# ===========================================

# Core LangChain
langchain==0.3.9
langchain-community==0.3.9
langchain-ollama==0.2.2

# Vector Store
chromadb==0.5.23

# Document Processing
pypdf==5.1.0

# Web UI
streamlit==1.40.1

# Configuration
python-dotenv==1.0.1

# Utilities (optional but recommended)
# tiktoken==0.7.0  # For token counting
# watchdog==4.0.0  # For file watching
